transcriber/requirements.txt
flask==2.0.1
werkzeug==2.0.2
openai-whisper==20231117
torch==2.0.1
numpy==1.24.3
soundfile==0.12.1
requests==2.28.2
gunicorn==20.1.0